requirements/development.txt
picamera
fabric
Pillow
numpy
nose